store counts as float

+++ heatmap.go
@@ -3,18 +3,18 @@ package main
 import "errors"
 
 type HeatMap struct {
-	GridLevel    int     `json:"gridLevel"`
-	Columns      int     `json:"columns"`
-	Rows         int     `json:"rows"`
-	MinX         float64 `json:"minX"`
-	MinY         float64 `json:"minY"`
-	MaxX         float64 `json:"maxX"`
-	MaxY         float64 `json:"maxY"`
-	CountsInts2D [][]int `json:"counts_ints2D"`
+	GridLevel    int         `json:"gridLevel"`
+	Columns      int         `json:"columns"`
+	Rows         int         `json:"rows"`
+	MinX         float64     `json:"minX"`
+	MinY         float64     `json:"minY"`
+	MaxX         float64     `json:"maxX"`
+	MaxY         float64     `json:"maxY"`
+	CountsInts2D [][]float64 `json:"counts_ints2D"`
 }
 
-func (h *HeatMap) max() int {
-	max := 0
+func (h *HeatMap) max() float64 {
+	max := 0.0
 	for _, row := range h.CountsInts2D {
 		for _, col := range row {
 			if max < col {
@@ -33,7 +33,7 @@ func (h *HeatMap) Add(other *HeatMap) error {
 	for y := 0; y < h.Rows; y++ {
 		// nullなら
 		if len(h.CountsInts2D[y]) == 0 {
-			h.CountsInts2D[y] = make([]int, h.Columns)
+			h.CountsInts2D[y] = make([]float64, h.Columns)
 		}
 		if len(other.CountsInts2D[y]) == 0 {
 			continue
